+++ packages/adapter-cloudflare-workers/index.js
@@ -62,7 +62,7 @@ export default function ({ config = 'wrangler.toml' } = {}) {
 				entryPoints: [`${tmp}/entry.js`],
 				outfile: main,
 				bundle: true,
-				external: ['__STATIC_CONTENT_MANIFEST'],
+				external: ['__STATIC_CONTENT_MANIFEST', 'cloudflare:*'],
 				format: 'esm',
 				loader: {
 					'.wasm': 'copy'