+++ packages/kit/src/core/postbuild/analyse.js
@@ -60,7 +60,7 @@ async function analyse({ manifest_path, env }) {
 		const node = await loader();
 
 		metadata.nodes[node.index] = {
-			has_server_load: node.server?.load !== undefined
+			has_server_load: node.server?.load !== undefined || node.server?.trailingSlash !== undefined
 		};
 	}
 
@@ -69,7 +69,7 @@ async function analyse({ manifest_path, env }) {
 		/** @type {Array<'GET' | 'POST'>} */
 		const page_methods = [];
 
-		/** @type {import('types').HttpMethod[]} */
+		/** @type {(import('types').HttpMethod | '*')[]} */
 		const api_methods = [];
 
 		/** @type {import('types').PrerenderOption | undefined} */
@@ -96,6 +96,8 @@ async function analyse({ manifest_path, env }) {
 			Object.values(mod).forEach((/** @type {import('types').HttpMethod} */ method) => {
 				if (mod[method] && ENDPOINT_METHODS.has(method)) {
 					api_methods.push(method);
+				} else if (mod.fallback) {
+					api_methods.push('*');
 				}
 			});