+++ packages/kit/src/runtime/server/utils.js
@@ -99,15 +99,7 @@ export async function handle_error_and_jsonify(event, options, error) {
 		return error.body;
 	} else {
 		if (__SVELTEKIT_DEV__ && typeof error == 'object') {
-			error = new Proxy(error, {
-				get: (target, property) => {
-					if (property === 'stack') {
-						return fix_stack_trace(target.stack);
-					}
-
-					return Reflect.get(target, property, target);
-				}
-			});
+			fix_stack_trace(error);
 		}
 
 		return (